% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saphana.R
\name{hana_conn}
\alias{hana_conn}
\title{获取hana数据库链接}
\usage{
hana_conn(
  ip = "192.168.2.160",
  port = "31013",
  db_name = "WP1",
  user_name = "HD_REMOTE",
  pwd = "Qq1234567890"
)
}
\arguments{
\item{ip}{地址}

\item{port}{端口号}

\item{db_name}{数据库名称}

\item{user_name}{用户名}

\item{pwd}{密码}
}
\value{
返回值
}
\description{
获取hana数据库链接
}
\examples{
hana_conn()
}
