% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saphana.R
\name{hana_select}
\alias{hana_select}
\title{查询数据}
\usage{
hana_select(var_hana_conn = hana_conn(), sql_str)
}
\arguments{
\item{var_hana_conn}{连接}

\item{sql_str}{SQL表达式}
}
\value{
返回结果
}
\description{
查询数据
}
\examples{
hana_select()
}
