% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saphana.R
\name{hana_select_paging}
\alias{hana_select_paging}
\title{分页查询}
\usage{
hana_select_paging(
  var_hana_conn = hana_conn(),
  sql_str,
  page_by = "fid",
  from = 1,
  to = 10000
)
}
\arguments{
\item{var_hana_conn}{连接}

\item{sql_str}{基本语句}

\item{page_by}{分页字段}

\item{from}{开始}

\item{to}{结束}
}
\value{
返回值
}
\description{
分页查询
}
\examples{
()
hana_select_paging
}
