% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saphana.R
\name{hana_syncData}
\alias{hana_syncData}
\title{同步sap hana数据}
\usage{
hana_syncData(
  var_hana_conn = hana_conn(),
  sql_str,
  table_name_src,
  table_name_target,
  page_by = "fid",
  query_pageCount = 1e+05,
  upload_pageCount = 5000,
  sleep_second = 3
)
}
\arguments{
\item{var_hana_conn}{连接}

\item{sql_str}{基本SQL}

\item{table_name_src}{来源表}

\item{table_name_target}{目标表}

\item{page_by}{分页字段}

\item{query_pageCount}{查询批次}

\item{upload_pageCount}{写入批次}

\item{sleep_second}{写入间隔}
}
\value{
返回值
}
\description{
同步sap hana数据
}
\examples{
hana_syncData()
}
