% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saphana.R
\name{hana_tableCount}
\alias{hana_tableCount}
\title{返回表的行数}
\usage{
hana_tableCount(var_hana_conn = hana_conn(), table_name)
}
\arguments{
\item{var_hana_conn}{连接}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
返回表的行数
}
\examples{
hana_tableCount()
}
