% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saphana.R
\name{hana_truncateTable}
\alias{hana_truncateTable}
\title{清空hana表中所有记录}
\usage{
hana_truncateTable(var_hana_conn = hana_conn(), table_name = "t_nka_fact")
}
\arguments{
\item{var_hana_conn}{连接}

\item{table_name}{表名}
}
\value{
无
}
\description{
清空hana表中所有记录
}
\examples{
hana_truncTable()
}
