% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saphana.R
\name{hana_update}
\alias{hana_update}
\title{测试数据数据}
\usage{
hana_update(var_hana_conn = hana_conn(), sql_str)
}
\arguments{
\item{var_hana_conn}{连接}

\item{sql_str}{字符串}
}
\value{
返回值
}
\description{
测试数据数据
}
\examples{
hana_update()
}
