% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saphana.R
\name{hana_upload_paging}
\alias{hana_upload_paging}
\title{hana数据上传}
\usage{
hana_upload_paging(
  var_hana_conn = hana_conn(),
  data,
  table_name,
  page_count = 5000,
  sleep_second = 3
)
}
\arguments{
\item{var_hana_conn}{连接}

\item{data}{数据}

\item{table_name}{表名}

\item{page_count}{页面}

\item{sleep_second}{间隔时间}
}
\value{
无返回值
}
\description{
hana数据上传
}
\examples{
hana_upload()
}
