% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saphana.R
\name{hana_writeTable}
\alias{hana_writeTable}
\title{hana写入数据库}
\usage{
hana_writeTable(
  var_hana_conn = hana_conn(),
  table_name,
  r_object,
  append = FALSE
)
}
\arguments{
\item{var_hana_conn}{连接}

\item{table_name}{表名}

\item{r_object}{对话}

\item{append}{是否附加}
}
\value{
返回值
}
\description{
hana写入数据库
}
\examples{
hana_writeTable()
}
